:lipstick: Apple Button 구성 및 컨벤션 정리
